Implemented and tested all file-system service operations.
Had to use a localized version of the "SDClass" from SD.h in order
to expose the lower level functionality available in raw SdFile
objects.  In the end, the code had to be copied (ugh) rather than
inherited as SDClass defines its state variables as private and these
state variables are needed to extend the class.